+++ scalameta/testkit/src/main/scala/scala/meta/testkit/CorpusFile.scala
@@ -8,7 +8,7 @@ import java.io.File
   * @param projectUrl The url of the Github project containing this source file.
   * @param commit The commit has at where this ScalaFile originates from.
   */
-case class ScalaFile(filename: String, projectUrl: String, commit: String) {
+case class CorpusFile(filename: String, projectUrl: String, commit: String) {
 
   /** Github "raw" url for this ScalaFile */
   def rawUrl: String = {
